Fix treasuryBecomesDeletableAfterTokenDelete EET
