+++ test-clients/src/main/java/com/hedera/services/bdd/suites/token/TokenAssociationSpecs.java
@@ -29,7 +29,6 @@ import com.hedera.services.bdd.spec.infrastructure.meta.ContractResources;
 import com.hedera.services.bdd.spec.transactions.token.TokenMovement;
 import com.hedera.services.bdd.suites.HapiApiSuite;
 import com.hederahashgraph.api.proto.java.AccountAmount;
-import com.hederahashgraph.api.proto.java.TokenSupplyType;
 import com.hederahashgraph.api.proto.java.TokenTransferList;
 import com.hederahashgraph.api.proto.java.TokenType;
 import org.apache.logging.log4j.LogManager;
@@ -38,7 +37,6 @@ import org.apache.logging.log4j.Logger;
 import java.nio.charset.StandardCharsets;
 import java.util.Collections;
 import java.util.List;
-import java.util.Map;
 import java.util.concurrent.atomic.AtomicLong;
 
 import static com.hedera.services.bdd.spec.HapiApiSpec.defaultHapiSpec;
@@ -54,7 +52,6 @@ import static com.hedera.services.bdd.spec.transactions.TxnVerbs.contractCreate;
 import static com.hedera.services.bdd.spec.transactions.TxnVerbs.cryptoCreate;
 import static com.hedera.services.bdd.spec.transactions.TxnVerbs.cryptoTransfer;
 import static com.hedera.services.bdd.spec.transactions.TxnVerbs.fileCreate;
-import static com.hedera.services.bdd.spec.transactions.TxnVerbs.fileUpdate;
 import static com.hedera.services.bdd.spec.transactions.TxnVerbs.mintToken;
 import static com.hedera.services.bdd.spec.transactions.TxnVerbs.tokenAssociate;
 import static com.hedera.services.bdd.spec.transactions.TxnVerbs.tokenCreate;
@@ -72,8 +69,6 @@ import static com.hederahashgraph.api.proto.java.ResponseCodeEnum.ACCOUNT_FROZEN
 import static com.hederahashgraph.api.proto.java.ResponseCodeEnum.ACCOUNT_IS_TREASURY;
 import static com.hederahashgraph.api.proto.java.ResponseCodeEnum.INVALID_SIGNATURE;
 import static com.hederahashgraph.api.proto.java.ResponseCodeEnum.INVALID_TOKEN_ID;
-import static com.hederahashgraph.api.proto.java.ResponseCodeEnum.REQUESTED_NUM_AUTOMATIC_ASSOCIATIONS_EXCEEDS_ASSOCIATION_LIMIT;
-import static com.hederahashgraph.api.proto.java.ResponseCodeEnum.TOKENS_PER_ACCOUNT_LIMIT_EXCEEDED;
 import static com.hederahashgraph.api.proto.java.ResponseCodeEnum.TOKEN_NOT_ASSOCIATED_TO_ACCOUNT;
 import static com.hederahashgraph.api.proto.java.ResponseCodeEnum.TRANSACTION_REQUIRES_ZERO_TOKEN_BALANCES;
 import static com.hederahashgraph.api.proto.java.TokenFreezeStatus.FreezeNotApplicable;
@@ -115,7 +110,6 @@ public class TokenAssociationSpecs extends HapiApiSuite {
 						dissociateHasExpectedSemanticsForDeletedTokens(),
 						dissociateHasExpectedSemanticsForDissociatedContracts(),
 						canDissociateFromDeletedTokenWithAlreadyDissociatedTreasury(),
-						cannotAssociateMoreThanTheLimit()
 				}
 		);
 	}
@@ -123,80 +117,6 @@ public class TokenAssociationSpecs extends HapiApiSuite {
 	@Override
 	public boolean canRunAsync() {
 		return true;
-	}
-
-	private HapiApiSpec cannotAssociateMoreThanTheLimit() {
-		final String treasury1 = "treasury1";
-		final String treasury2 = "treasury2";
-		final String user1 = "user1";
-		final String user2 = "user2";
-		final String key = "key";
-		final String token1 = "token1";
-		final String token2 = "token2";
-		final String token3 = "token3";
-		return defaultHapiSpec("CannotAssociateMoreThanTheLimit")
-				.given(
-						newKeyNamed(key),
-						cryptoCreate(treasury1),
-						cryptoCreate(treasury2),
-						tokenCreate(token1)
-								.supplyType(TokenSupplyType.FINITE)
-								.tokenType(TokenType.FUNGIBLE_COMMON)
-								.treasury(treasury1)
-								.maxSupply(500)
-								.kycKey(key)
-								.initialSupply(100),
-						tokenCreate(token2)
-								.supplyType(TokenSupplyType.FINITE)
-								.tokenType(TokenType.FUNGIBLE_COMMON)
-								.treasury(treasury2)
-								.maxSupply(500)
-								.kycKey(key)
-								.initialSupply(100),
-						fileUpdate(APP_PROPERTIES)
-								.payingWith(ADDRESS_BOOK_CONTROL)
-								.overridingProps(Map.of("tokens.maxPerAccount", "0"))
-				)
-				.when(
-						tokenCreate(token3)
-								.supplyType(TokenSupplyType.FINITE)
-								.tokenType(TokenType.FUNGIBLE_COMMON)
-								.treasury(treasury2)
-								.maxSupply(500)
-								.kycKey(key)
-								.initialSupply(100)
-								.hasKnownStatus(TOKENS_PER_ACCOUNT_LIMIT_EXCEEDED),
-						cryptoCreate(user1)
-								.maxAutomaticTokenAssociations(1)
-								.hasPrecheck(REQUESTED_NUM_AUTOMATIC_ASSOCIATIONS_EXCEEDS_ASSOCIATION_LIMIT),
-						cryptoCreate(user1),
-						tokenAssociate(user1, token1)
-								.hasKnownStatus(TOKENS_PER_ACCOUNT_LIMIT_EXCEEDED),
-						tokenAssociate(treasury1, token2)
-								.hasKnownStatus(TOKENS_PER_ACCOUNT_LIMIT_EXCEEDED)
-				)
-				.then(
-						fileUpdate(APP_PROPERTIES)
-								.payingWith(ADDRESS_BOOK_CONTROL)
-								.overridingProps(Map.of("accounts.limitTokenAssociations", "false")),
-						tokenCreate(token3)
-								.supplyType(TokenSupplyType.FINITE)
-								.tokenType(TokenType.FUNGIBLE_COMMON)
-								.treasury(treasury2)
-								.maxSupply(500)
-								.kycKey(key)
-								.initialSupply(100),
-						cryptoCreate(user2)
-								.maxAutomaticTokenAssociations(10_000),
-						tokenAssociate(user1, token1),
-						tokenAssociate(treasury1, token2),
-						fileUpdate(APP_PROPERTIES)
-								.payingWith(ADDRESS_BOOK_CONTROL)
-								.overridingProps(Map.of("tokens.maxPerAccount", "1000")),
-						fileUpdate(APP_PROPERTIES)
-								.payingWith(ADDRESS_BOOK_CONTROL)
-								.overridingProps(Map.of("accounts.limitTokenAssociations", "true"))
-				);
 	}
 
 	public HapiApiSpec handlesUseOfDefaultTokenId() {
