+++ test-clients/src/main/java/com/hedera/services/bdd/suites/token/TokenManagementSpecs.java
@@ -77,7 +77,7 @@ public class TokenManagementSpecs extends HapiApiSuite {
 	private static final Logger log = LogManager.getLogger(TokenManagementSpecs.class);
 
 	public static void main(String... args) {
-		new TokenManagementSpecs().runSuiteSync();
+		new TokenManagementSpecs().runSuiteAsync();
 	}
 
 	@Override
